Better search functionalities
The search function now redisters the likeness of the search input to all the columns in the table for articles. It also displays a message when the search didn't find ny matching results.

Furthermore, two more javascript functions have been added and the old function for searching has been made redundant and was therefore removed.