Final touch up
Changed the amount of articles published and queries performed to 200 each. Also added an interval to the loop for the autoSearch() to make sure the browser can keep up with the script.

Also changed the variable $log_time so i stores the response time in the form of JSON objects.