Final commit to test
The finished files that were used to execute the experiment. Added a delete method, for time-efficiency. Redis.php and Memcached.php are base on servers.php and each is used for their respective cache technology.

The testfiles are used to test the connections to the cache servers. The phpinfo file is used to verify the installations of extensions in Xampp